full stack open - part 4: exercise 4.21 blog list expansions, step 9

+++ 07.full-stack-open-2023/part4/blog-list/controllers/blogs.js
@@ -33,6 +33,18 @@ blogsRouter.post('/', async (request, response) => {
 blogsRouter.delete('/:id', async (request, response) => {
     const id = request.params.id
 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'invalid token' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(id)
+
+    if (user._id.toString() !== blog.user.toString()) {
+        return response.status(401).json({ error: 'invalid token' })
+    }
+
     const removedBlog = await Blog.findByIdAndRemove(id)
 
     if (removedBlog)
